+++ Features/Analytics/GetAdminDashboard/Queries/GetAllDiplomasNubmerQuery.cs
@@ -2,6 +2,6 @@
 
 namespace exam_system.Features.Analytics.GetAdminDashboard.Queries
 {
-    public record GetAllDiplomasNubmerQuery : IRequest<int>;
+    public record GetAllDiplomasNubmerQuery : IRequest<double>;
     
 }
