+++ Features/Analytics/GetAdminDashboard/Queries/GetAllQuizesNumberQuery.cs
@@ -2,6 +2,6 @@
 
 namespace exam_system.Features.Analytics.GetAdminDashboard.Queries
 {
-    public record GetAllQuizesNumberQuery : IRequest<int>;
+    public record GetAllQuizesNumberQuery : IRequest<double>;
     
 }
