+++ Features/Analytics/GetAdminDashboard/Queries/GetAveragePassRateQuery.cs
@@ -2,6 +2,6 @@
 
 namespace exam_system.Features.Analytics.GetAdminDashboard.Queries
 {
-    public record GetAveragePassRateQuery : IRequest<float>;
+    public record GetAveragePassRateQuery : IRequest<double>;
     
 }
