+++ Features/Shared/AdminDashoardDto.cs
@@ -2,12 +2,12 @@
 {
     public class AdminDashoardDto
     {
-        public int TotalRegistredUsers { get; set; }
-        public int TotalActiveUsers { get; set; }
-        public int TotalNumberOfDiplomas { get; set; }
-        public int TotalNumberOfQuizes { get; set; }
-        public int TotalAttembts { get; set; }
-        public float AveragePassRate { get; set; }
+        public double TotalRegistredUsers { get; set; }
+        public double TotalActiveUsers { get; set; }
+        public double TotalNumberOfDiplomas { get; set; }
+        public double TotalNumberOfQuizes { get; set; }
+        public double TotalAttembts { get; set; }
+        public double AveragePassRate { get; set; }
 
     }
 }
